+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+- 3.1.0
+    - Implemented the singular value decomposition (SVD) in the pure-PHP library
+    - Implemented the Moore-Penrose pseudoinverse in the pure-PHP library via SVD
+
 - 3.0.11
     - Update to Zephir 1.5.0
     - Fixed RREF/REF pivot float tolerance